requirements.txt
python-binance
mariadb
pandas
pandas-ta
ecs-logging